+++ scripts/addRedirects.cjs
@@ -7,7 +7,7 @@ const { JSDOM } = require("jsdom");
 const crawl = require("./crawl.cjs");
 
 const APP_HTML_LOCATION = resolve(__dirname, "../src/app.html");
-const BUILD_DIRECTORY = resolve(__dirname, "../build");
+const OUT_DIRECTORY = resolve(__dirname, "../static");
 const ROUTES_DIRECTORY = resolve(__dirname, "../src/routes");
 
 if (!existsSync(APP_HTML_LOCATION)) {
@@ -16,9 +16,9 @@ if (!existsSync(APP_HTML_LOCATION)) {
   );
 }
 
-if (!existsSync(BUILD_DIRECTORY)) {
+if (!existsSync(OUT_DIRECTORY)) {
   throw new Error(
-    `[addRedirects] | Build directory at ${BUILD_DIRECTORY} does not exist`
+    `[addRedirects] | Build directory at ${OUT_DIRECTORY} does not exist`
   );
 }
 
@@ -60,5 +60,5 @@ for (const route of routes) {
 
   const serialised = root.serialize();
 
-  writeFileSync(join(BUILD_DIRECTORY, `${route}.html`), serialised);
+  writeFileSync(join(OUT_DIRECTORY, `${route}.html`), serialised);
 }